# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Asset Management System API (ISO 27001)
  version: 1.0.0
  description: RESTful API for IT Asset & ISMS Management with ISO 27001 compliance
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: Development
paths:
...
  /assets/types/{typeId}/subtypes:
    post:
      summary: Create asset subtype for inventory-aware asset classification
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetSubtypeInput'
      responses:
        '201': { description: Asset subtype created }
  /assets/inventory/preview:
    get:
      summary: Preview next inventory number from asset type or subtype pattern
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: assetTypeId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: assetSubtypeId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Inventory preview with enabled flag and nextInventoryNumber }

  # ==================== Assets ====================
  /assets:
    get:
      summary: List assets (paginated, filterable)
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: assetTypeId
          in: query
          schema: { type: string }
        - name: criticality
          in: query
          schema: { type: string, enum: [low, medium, high, critical] }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create asset
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
  /assets/{id}:
    get:
      summary: Get asset by ID
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
...
  /controls/tests:
    post:
      summary: Create a control implementation test with generic evidence links
      tags: [Controls]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlTestInput'
      responses:
        '201': { description: Control test created }
  /assets/{id}/graph:
    get:
      summary: Get asset dependency graph (centered on this asset)
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: depth
          in: query
          schema: { type: integer, default: 3 }
  /assets/graph:
    get:
      summary: Get full asset dependency graph
      tags: [Assets]
      security: [{ bearerAuth: [] }]
  /assets/{id}/impact-analysis:
    get:
      summary: Impact analysis for asset failure scenario
      tags: [Assets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }

  # ==================== Risks ====================
...
  /audit-logs:
    get:
      summary: List audit log entries (read-only)
      tags: [Audit]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: objectType
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }

  # ==================== Contracts ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ControlTestInput:
      type: object
      required: [controlImplementationId, testType, testedBy, result]
      properties:
        controlImplementationId: { type: string, format: uuid }
        testType: { type: string }
        testMethod: { type: string }
        testedBy: { type: string }
        testedAt: { type: string, format: date-time }
        result: { type: string }
        effectivenessRating: { type: integer, minimum: 0, maximum: 100 }
        findings: { type: string }
        evidenceRequired: { type: boolean }
        nextTestDate: { type: string, format: date-time }
        evidenceLinks:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLinkInput'
    EvidenceLinkInput:
      type: object
      required: [evidenceId]
      properties:
        evidenceId: { type: string, format: uuid }
        relationType: { type: string, default: supports }
...
    AssetSubtypeInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        inventoryEnabled: { type: boolean, nullable: true }
        inventoryPattern: { type: string, nullable: true }
...
    AssetInput:
      type: object
      required: [name, assetTypeId]
      properties:
        name: { type: string }
        description: { type: string }
        assetTypeId: { type: string, format: uuid }
        assetSubtypeId: { type: string, format: uuid }
        inventoryNumber: { type: string, description: Optional manual number; otherwise generated from subtype/type inventory pattern when enabled. }
        criticality: { type: string, enum: [low, medium, high, critical] }
        networkAddresses:
          type: array
          items:
            type: object
            properties:
              address: { type: string }
              type: { type: string, enum: [ipv4, ipv6, cidr, hostname] }
              primary: { type: boolean }
    AssetResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayId: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        assetTypeId: { type: string }
        assetSubtypeId: { type: string, nullable: true }
        inventoryNumber: { type: string, nullable: true }
        criticality: { type: string }
        lifecycleStatus: { type: string }
        confidentialityNeed: { type: string }
        integrityNeed: { type: string }
        availabilityNeed: { type: string }
    PaginatedResponse:
      type: object
      properties:
        data: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }

# ==================== Phase 8: API, Betrieb und Nachweise ====================

  # ==================== Health & System (Phase 8) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== Phase 8 Schemas ====================
...